Fix compiler warnings

+++ sjasm/io_trd.cpp
@@ -32,6 +32,7 @@
 #include <fstream>
 #include <numeric>
 #include <vector>
+#include <cstdint>
 
 namespace {
 
@@ -380,7 +381,8 @@ int TRD_AddFile(const Filename& fname, const HobetaFilename& fhobname, int start
 	}
     image.DeleteFile(fhobname);
 
-    const unsigned char autostartData[] = {0x80, 0xaa, autostart & 0xff, autostart >> 8};
+    const unsigned char autostartData[] =
+            { 0x80, 0xaa, (uint8_t)(autostart & 0xff), (uint8_t)(autostart >> 8) };
 
     const unsigned realLength = autostart > 0 ? realLength + sizeof(autostartData) : length;
     if (void* data = image.AddFile(fhobname, IsBasic(fhobname) ? length : start, realLength)) {
